NewsAggregator
This is a NewsAggregator application, will display the aggregated news from a news website using web scraping. In this project, Flask is used as the web framework, Python library - BeautifulSoup is used for scraping HTML data from a news website and SQLAlchemy - Object Relational Mapper (ORM) is used for handling the Sqlite database.